requirements.txt
google-cloud-bigquery
pandas
pyarrow
google-cloud-bigquery-storage
pandas-gbq
xmltodict